Renamed TestCases to matching with the JUnit class name regex.
