+++ src/test/java/club/psychose/testsuite/ibo/testcases/datatypes/TestUInt16DataType.java
@@ -41,7 +41,7 @@ import java.nio.ByteOrder;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public final class TC0004DatatypeUInt16 {
+public final class TestUInt16DataType {
     @Test
     public void executeTestCase () {
         // Out-of-Bounds Test.
