+++ src/test/java/club/psychose/testsuite/ibo/testcases/datatypes/TestUInt24DataType.java
@@ -40,7 +40,7 @@ import java.nio.ByteOrder;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class TC0011DatatypeUInt24 {
+public class TestUInt24DataType {
     @Test
     public void executeTestCase () {
         // Out-of-Bounds Test.
